# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: InstaVEG - Hyperlocal grocery marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sellerId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (seller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/trending:
    get:
      operationId: getTrendingProducts
      tags: [products]
      summary: Get trending products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Trending products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/fresh-picks:
    get:
      operationId: getFreshPicks
      tags: [products]
      summary: Get fresh picks (recently added)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Fresh picks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (seller/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (seller/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # REVIEWS
  /products/{id}/reviews:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # CART
...
  /sellers:
    get:
      operationId: listSellers
      tags: [sellers]
      summary: List nearby/all sellers
      parameters:
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sellers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Seller"
  /sellers/{id}:
    get:
      operationId: getSeller
      tags: [sellers]
      summary: Get seller by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Seller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Seller"
...
  /sellers/dashboard/orders:
    get:
      operationId: getSellerOrders
      tags: [sellers]
      summary: Get orders for seller dashboard
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Seller orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /sellers/profile:
    patch:
      operationId: updateSellerProfile
      tags: [sellers]
      summary: Update seller profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellerProfileUpdate"
      responses:
        "200":
          description: Updated seller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Seller"

  # ADMIN
...
components:
  schemas:
...
    Seller:
      type: object
      required: [id, userId, storeName, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        storeName:
          type: string
        description:
          type: ["string", "null"]
        logo:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        deliveryRadius:
          type: ["number", "null"]
        rating:
          type: ["number", "null"]
        totalOrders:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        createdAt:
          type: string
          format: date-time
        ownerName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isVerified:
          type: boolean
    SellerProfileUpdate:
      type: object
      properties:
        storeName:
          type: string
        description:
          type: string
        logo:
          type: string
        address:
          type: string
        city:
          type: string
        deliveryRadius:
          type: number
        phone:
          type: string
...
    Product:
      type: object
      required: [id, name, price, stock, sellerId, categoryId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        mrp:
          type: ["number", "null"]
        unit:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        sellerId:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        isFresh:
          type: boolean
        isOrganic:
          type: boolean
        nutritionInfo:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        sellerName:
          type: ["string", "null"]
        sellerCity:
          type: ["string", "null"]
    ProductInput:
      type: object
      required: [name, price, stock, categoryId, unit]
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
        price:
          type: number
        mrp:
          type: ["number", "null"]
        unit:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        categoryId:
          type: integer
        isFresh:
          type: boolean
        isOrganic:
          type: boolean
        nutritionInfo:
          type: string
        tags:
          type: array
          items:
            type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        mrp:
          type: ["number", "null"]
        unit:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        categoryId:
          type: integer
        isFresh:
          type: boolean
        isOrganic:
          type: boolean
        nutritionInfo:
          type: string
        tags:
          type: array
          items:
            type: string
    ProductList:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # CART
...
    Order:
      type: object
      required: [id, userId, status, paymentStatus, total, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        sellerId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, confirmed, preparing, dispatched, delivered, cancelled]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        paymentMethod:
          type: ["string", "null"]
        razorpayOrderId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        addressId:
          type: ["integer", "null"]
        addressSnapshot:
          type: ["string", "null"]
        subtotal:
          type: number
        deliveryFee:
          type: number
        platformFee:
          type: number
        total:
          type: number
        eta:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        sellerName:
          type: ["string", "null"]
        buyerName:
          type: ["string", "null"]
    OrderItem:
      type: object
      required: [id, productId, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        productName:
          type: ["string", "null"]
        productImage:
          type: ["string", "null"]
        productUnit:
          type: ["string", "null"]
...
    Review:
      type: object
      required: [id, productId, userId, rating, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        userAvatar:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    # ADMIN
...