# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InstaVEG - Hyperlocal grocery marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 2 */
  name: string;
  description?: string;
  price: number;
  /** @nullable */
  mrp?: number | null;
  unit: string;
  images?: string[];
  stock: number;
  categoryId: number;
  isFresh?: boolean;
  isOrganic?: boolean;
  nutritionInfo?: string;
  tags?: string[];
}